Use only a single Playground button
