+++ packages/fiberplane-hono/src/constants.ts
@@ -1,5 +1,4 @@
 export const DEFAULT_PLAYGROUND_SERVICES_URL = "https://services.fp.dev";
-// export const DEFAULT_PLAYGROUND_SERVICES_URL = "http://localhost:7676";
 
 /**
  * Fiberplane services URL.
