+++ packages/playground/src/components/WorkflowDetail/StepDetail/index.ts
@@ -1,0 +1,1 @@
+export { StepDetail } from "./StepDetail";
